fix uint8 Content.SetLevel issue

+++ writescript.go
@@ -90,7 +90,7 @@ func (w *WriteScript) Process(plugin, data string, headerOn bool) error {
 	vm.Set("setLevel", func(call otto.FunctionCall) otto.Value {
 		val, err := call.Argument(0).ToInteger()
 		if err == nil {
-			w.Content.SetLevel(int(val))
+			w.Content.SetLevel(uint8(val))
 		}
 		return otto.Value{}
 	})
